update component Products

+++ src/components/Products/Product/index.js
@@ -1,6 +1,6 @@
 import React from "react";
-import { useDispatch } from "react-redux";
-import { Link, useHistory } from "react-router-dom";
+import { useDispatch, useSelector } from "react-redux";
+import { useHistory } from "react-router-dom";
 import Swal from "sweetalert2";
 
 import {
@@ -10,26 +10,32 @@ import {
 
 const Product = ({ product }) => {
   const { _id, name, price, detail, category } = product;
-
   const dispatch = useDispatch();
   const history = useHistory(); // Habilitar history para redireccionar.
 
+  const user = useSelector((state) => state.authState);
   const onDeleteProduct = (id) => {
-    // preguntar al usuario
-    Swal.fire({
-      title: "¿Estas seguro?",
-      text: "Esta accion es irreversible.",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#3085d6",
-      cancelButtonColor: "#d33",
-      confirmButtonText: "Aceptar",
-      cancelButtonText: "Cancelar",
-    }).then((result) => {
-      if (result.value) {
-        dispatch(deleteProductAction(id));
-      }
-    });
+    // Is login
+
+    if (user.email) {
+      // preguntar al usuario
+      Swal.fire({
+        title: "¿Estas seguro?",
+        text: "Esta accion es irreversible.",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Aceptar",
+        cancelButtonText: "Cancelar",
+      }).then((result) => {
+        if (result.value) {
+          dispatch(deleteProductAction(id));
+        }
+      });
+    } else {
+      history.push(`/users/login`);
+    }
   };
 
   const onEditRedirection = (id) => {
